% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{datasus_idade_sexo_dose}
\alias{datasus_idade_sexo_dose}
\title{datasus_idade_sexo_dose}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 417 rows and 4 columns.
}
\usage{
datasus_idade_sexo_dose
}
\description{
datasus_idade_sexo_dose
}
\keyword{datasets}
