% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{pop_idade_sexo}
\alias{pop_idade_sexo}
\title{pop_idade_sexo}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 224 rows and 3 columns.
}
\usage{
pop_idade_sexo
}
\description{
pop_idade_sexo
}
\keyword{datasets}
